+++ src/modules/display.js
@@ -1,0 +1,7 @@
+(function () {
+  function renderMarkdown(element, markdownToHtml, markdown) {
+    element.innerHTML = markdownToHtml(markdown || "");
+  }
+
+  window.MDBasicsDisplay = { renderMarkdown };
+})();